DO NOT PUSH: Serperated reddit api into serperate files. Need to rework the saving of pictures. (currently broken)
