+++ Tokens
@@ -1,1 +1,1 @@
-{"access_token":"456556608586-_4J4wn-8jlgDERVkxl7Vd1hrSRCOkg","token_type":"bearer","expires_in":3600,"refresh_token":"456556608586-94KBHNR3e9umrj8xhUUTzsfSJIvi0A","scope":"edit flair history identity modconfig modflair modlog modposts modwiki mysubreddits privatemessages read report save submit subscribe vote wikiedit wikiread"}
+{"access_token":"456556608586-pWzAV1k3EY9Gx5xdgQqoCR8CzNdSxA","token_type":"bearer","expires_in":3600,"refresh_token":"456556608586-4CavVYdmOjjt0jEY93g_P8W0JidukQ","scope":"edit flair history identity modconfig modflair modlog modposts modwiki mysubreddits privatemessages read report save submit subscribe vote wikiedit wikiread"}
