+++ package.json
@@ -1,9 +1,1 @@
-{  
-  "dependencies": {
-    "axios": "^0.21.1",
-    "endb": "^0.23.0",
-    "express": "^4.17.1",
-    "nedb": "^1.8.0",
-    "nodemon": "^2.0.7"
-  }
-}
+{"dependencies":{"axios":"^0.21.1","endb":"^0.23.0","express":"^4.17.1","mkdirp":"^0.5.5","nedb":"^1.8.0","nodemon":"^2.0.7"}}
